feat(trusty-mpm): add tm-circuit-breaker, tm-verification-protocols, tm-tool-usage-guide, tm-git-file-tracking, tm-adr skills
First batch of the /tm- skills portfolio: five KEEP-FULL skills rewritten
tm-native from their mpm-* sources, grounded in the actual PM_INSTRUCTIONS.md
canonical circuit-breaker table (CB#1-10, CB#14), the real mcp__trusty-mpm__* /
mcp__trusty-memory__* / mcp__trusty-search__* / mcp__trusty-review__* tool
families, and this repo's real docs/adr/ convention.

Also splits daemon/doctor.rs into doctor.rs + doctor_fs_checks.rs (the four
filesystem-only probes) to stay under the 500-SLOC production cap after A2's
check_skill_source addition — mirrors the existing doctor_output_style.rs
split, no behavior change.

Part of the /tm- skills portfolio epic.
